Remove hard deletion for outlet model.
Outlets cannot be hard deleted because the celery task functions that
scrapes the blogs depends on it. To deal with the fact that someone may
want to stop fetching articles from a given source, the Outlet model was
modified by adding an `active` boolean field that prevents or not the
task function to run on it.

The API's docs was changed adding information on how a user can query
articles from the database.